pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Pau <plabartabajo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7.1"
pyodbc = "^4.0.32"
pandas = "^1.3.3"
python-dotenv = "^0.19.0"
tqdm = "^4.62.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
